Razor Pages with EF Core migrations in ASP.NET Core
1- Drop the database
2- Create an initial migration
3- Remove EnsureCreated

When a new app is developed, the data model changes frequently. Each time the model changes, the model gets out of sync with the database. This tutorial series started by configuring the Entity Framework to create the database if it doesn't exist. Each time the data model changes, the database needs to be dropped. The next time the app runs, the call to EnsureCreated re-creates the database to match the new data model. The DbInitializer class then runs to seed the new database.

This approach to keeping the DB in sync with the data model works well until the app needs to be deployed to production. When the app is running in production, it's usually storing data that needs to be maintained. The app can't start with a test DB each time a change is made (such as adding a new column). The EF Core Migrations feature solves this problem by enabling EF Core to update the DB schema instead of creating a new database.

Rather than dropping and recreating the database when the data model changes, migrations updates the schema and retains existing data.

Database migration should be done as part of deployment, and in a controlled way. Production database migration approaches include:
- Using migrations to create SQL scripts and using the SQL scripts in deployment.
- Running dotnet ef database update from a controlled environment.