Razor Pages with EF Core in ASP.NET Core - Read Related Data (IMPORTANT)
This tutorial shows how to read and display related data. Related data is data that EF Core loads into navigation properties.

Eager loading is specified with the Include and ThenInclude methods.

Separate queries with Load is more like explicit loading than eager loading.

Explicit loading with separate queries results in multiple queries sent to the database. With explicit loading, the code specifies the navigation properties to be loaded.

When the entity is first read, related data isn't retrieved. The first time a navigation property is accessed, the data required for that navigation property is automatically retrieved. A query is sent to the database each time a navigation property is accessed for the first time. Lazy loading can hurt performance, for example when developers use N+1 queries.

https://learn.microsoft.com/en-us/aspnet/core/data/ef-rp/read-related-data?view=aspnetcore-7.0&tabs=visual-studio

1- Create Course pages - Scaffold Course pages
1.1- Display the department name
1.2- Create CourseViewModel
1.3- Loading related data with Select

2- Create Instructor pages
2.1- Create a view model
2.2- Scaffold Instructor pages
2.3- Update Pages/Instructors/Index.cshtml.cs
2.4- Update Pages/Instructors/Index.cshtml